changes to move data to azure blob
