--- constants.py
@@ -1,2 +1,0 @@
-TARGET_FILE = r"data/daily_target.json"
-DOB_FILE = r"data/dob.json"
