+++ pyproject.toml
@@ -13,6 +13,7 @@ plotly = "^5.24.1"
 black = "^24.10.0"
 ruff = "^0.8.4"
 isort = "^5.13.2"
+azure-storage-blob = "^12.24.0"
 
 
 [build-system]
